browser: build environment stability improvements
Improve stability of `gulp watch` with respect to error recovery.

1) rearrange jscs to run once per file, and ensure that is is "plumbed"
with gulp-plumber.

2) improve logic of restarting the watch and rebuilding the entire app
so that the watch doesn't crash and so that full rebuilds are executed
on the next save after a build error.

3) node-sass is crashing the build when it is attempting to make source
maps and encounters a syntax error.  The crash is a segfault and there
is no opportunity to catch the error.  The process exits ungracefully.

As such, create buildParams.sassCompiler property to allow developer to
choose between the lesser of evils.  See comments in buildParam.js
for more information.

Also plumb gulp-sass "more better" so that its errors are handled
when they can be.

Warning: At present. the buildParams.sassCompiler = 'ruby-sass' option is
experimental and is not likely to work yet.

Note: The code changes in this commit reflect the general "draftiness"
of the build script. This script will be overhauled when time permits.

4) improvements to speed of build environment.