feature(explore): initial implementation
Enable the user (contributor or visitor) to search, filter and browse
question-and-answer documents.

Filters are designed to support "shadow queries".

Beginnings of the "view-qna-doc" feature -- support rendering of
documents and accurate vote handling. No dynamic features of this
state are yet included.

Limitations include:

* cannot filter by date range
* cannot filter by "resolved only"
* all tags dialog does not have the intended data and is presently
disconnected

Also includes

* a relatively major layout overhaul, moving aware from using bootstrap
to control layout.
* many various fixes and improvements