browser: error on invalid data from server
These changes are intended to ensure that only sessions with valid
metadata are engaged in the browser. Since we do not get back a session
id, there is no way to communicate to the middle tier that we are
rejecting a session.

TODO: fix the Java /v1/login endpoint (#224) and then verify proper
end-to-end functionality.

In mlModelBase,

* in assignData, use this.mergeData instead of mlUtil.merge, and
do not double-call this.testValidity

In mlAuth,

* in authenticate, reject if invalid data received
* in restoreSession, do not restore an invalid session

In login dialog,

* create a clean sesion object on rejection, copying user name, so that
future login attempts are allowed (i.e. we don't have a half-baked
session object
* remove runtime error message once the user starts to reattempt login
by entering password text