e2e: cleanstart.sh tweaks
Add messaging, exit on certain failures, fix bugs.

This script is only useful on OSX. It represents a path that is more
likely than some others to yield a working build and "clean start".

Note that while it won't work on other platforms, its alogorithm
should work and were it built into a Gradle task may be a decent
method of ensuring a clean installation.

At this time, dbTeardown is skipped in favor of **wiping out the
Marklogic/Data** directory, which means that if you run this on your
MarkLogic 8.0 installation, you will lose all of your data.