e2e/java: cleanstart: use dbTeardown instead of wiping out ML data
Now that server isn't sending HTML errors and gradle isn't choking on
them, we can use dbTeardown to wipe out only the Samplestack
configuration on the MarkLogic server instead of deleting the data
directory.

This makes for less upheaval in the user's environment.

Also clean up and add comments to the script.