*: deploy built app to java server
So that users who do not wish to develop the browser app may still use
it, deploy a built version of the application with each build from
the browser.

Create a rewriter class in Java to support HTML5 "push state" routing.
(This turned out to be just as easy as making it support "index.html"
as the default file, and causes the URLs not to be switched up between
Node and Java renderings.)