fix(e2e): handle "name" property of Page object constructor correctly
Page object constructors were using the "name" property to indicate the
"code-ish" name for them when they were added to World.

The "name" property of a constructor is off limits and so instead we
had the constructor's actual name being used.

Use "pageName" property to define pages' names in order to avoid this
conflict.