e2e: restore e2e test functionality, other post-skin fixes
This is a relatively wide-ranging commit. It:

* adds back many css classes that are needed as e2e test hooks
* fixes exit code from tests for travis
* makes Chrome the default browser for local tests
* removes ids from html so as not to confuse angular or ourselves
* changes expectations to match latest seed data
* a few other miscellaneous changes to re-enable tests
* fix presentation of error messages, spinners
* skips MarkLogic install when secure env vars not available
* disables e2e tests on travis when secure env vars not available

Note on Chrome being the default: it appears that something in the
new skin is tripping up PhantomJS. This is unfortunate, and we should
track it down. It may be PhantomJS itself, it may be the agedness of
it's Webkit, and css transitions are a likely culprit.

In the meantime, PhantomJS think the login button is invisible and
refuses to click it. As such, it can't really be used.

This makes headless testing *only available via SauceLabs, which in
turn means that we cannot e2e test PRs which cross repositories.