browser: Date picker wrap-up
Adds on blur and on enter key as trigger to apply picker date value while manually entering a date.

Add validation where if the currently focused date input is set to an invalid range, the other input is set to match its date, setting the range to a span of one month.

Closes #412 and Closes #414

Briefly swaps out the bound model value so when the pop-up renders, it reflects the placeholder date.

Closes #469 and closes #413