e2e: sudo apt-get update before MarkLogic on travis
Update packages lists before attempt to install MarkLogic and deps.

This is the last of several commits that improve testing
circumstances which came to a head in issue 481.

Note: IE tests may still fail due to slow travis, but these
commits minimize the changes that the test suite will hang.

Closes #481