browser: low-level "Page Unavailable"
This commit implements logic to support "Page Unavailble" requirements in #443.

It makes no particular effort to format/word properly the display.

As such, what is needed is to render the proper message.

The $rootScope.errorCondition property will either be falsy or in the
set ['preventedWithSession', 'preventedNoSession',
'notFoundWithSession', 'notFoundNoSession', 'authRequiredNav']

The first two represent a 401 or 403 from the REST server where it
has "errored" us as we tried to make an action request.

The 3rd and fourth represent a 404 from the REST server, and
the last is where the browser prevents navigation to a page when the
page is intended only for logged-in user ("ask" page).

The _layout.html template has logic as of this commmit to render only
one of those strings if they are set.

TODO: create formatted renderings of friendly error messages for these
conditions, all the user to "go back" to their last good state via
a click (it would just go back 1 in the history).