e2e: don't start selenium until middle-tier ready
Middle-tier startup in Spring takes much longer than selenium startup.

As such, and in particular in cloud environment, the selenium resources
are idle while the middle-tier is still going through its paces, and
when the selenium environment is OpenSauce, this is consuming a scarce
resource.

Alter task dependencies such that the selenium-start task depends on
the middle-tier-start task, and thus selenium start waits for successful
middle-tier start.