node: search qna docs, add custom snippets
Main update in this commit is custom snippeting (which is based on a
transform). There is also some general refactoring.

Currently supported in doc search:
- various qtext-based constraints described in Search Tips (e.g.,
"askedBy:joeUser")
- date range constraints
- tag constraints
- sorting
- pagination
- custom snippets that distinguish between question and answer origin
(same as in Java tier)

Supported but not well tested:
- "my contributions" constraint
- "resolved only" constraint
(both rely on auth to some extent and auth is broken in my env right
now)

A values call is always made to get min and max date bounds from search
results. This may not be needed when to/from dates are submitted from
browser.

Currently missing in the Node results payload (but present in the Java
results payload) is the qtext and query information. This is currently
not supported in the Node Client but I've requested it. Not sure if our
app actually uses this information