sass processing: change files base before sass
"gulp-tap"ing the files that come out of gulp-sass was rendering a
stream which cause event-stream.merge to fail.

Tap the inputs to the sass processing and change *their* base, so that
the *resulting* stream isn't touched by the tap plugin.

It is not known certainly, but this workaround may be necessitated
by what is discussed in geejs/gulp-tap#6